chore: separate icon and emoji icons
